+++ packages/std/src/api.rs
@@ -3,7 +3,7 @@
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
-use crate::errors::{Error, SystemError};
+use crate::errors::{StdError, SystemError};
 use crate::HumanAddr;
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
@@ -80,36 +80,36 @@ impl std::fmt::Display for ApiError {
     }
 }
 
-impl From<Error> for ApiError {
-    fn from(value: Error) -> Self {
+impl From<StdError> for ApiError {
+    fn from(value: StdError) -> Self {
         match value {
-            Error::Base64Err { source, .. } => ApiError::Base64Err {
+            StdError::Base64Err { source, .. } => ApiError::Base64Err {
                 source: format!("{}", source),
             },
-            Error::ContractErr { msg, .. } => ApiError::ContractErr {
+            StdError::ContractErr { msg, .. } => ApiError::ContractErr {
                 msg: msg.to_string(),
             },
-            Error::DynContractErr { msg, .. } => ApiError::DynContractErr { msg },
-            Error::NotFound { kind, .. } => ApiError::NotFound {
+            StdError::DynContractErr { msg, .. } => ApiError::DynContractErr { msg },
+            StdError::NotFound { kind, .. } => ApiError::NotFound {
                 kind: kind.to_string(),
             },
-            Error::NullPointer { .. } => ApiError::NullPointer {},
-            Error::ParseErr { kind, source, .. } => ApiError::ParseErr {
-                kind: kind.to_string(),
-                source: format!("{}", source),
-            },
-            Error::SerializeErr { kind, source, .. } => ApiError::SerializeErr {
+            StdError::NullPointer { .. } => ApiError::NullPointer {},
+            StdError::ParseErr { kind, source, .. } => ApiError::ParseErr {
                 kind: kind.to_string(),
                 source: format!("{}", source),
             },
-            Error::Utf8Err { source, .. } => ApiError::Utf8Err {
+            StdError::SerializeErr { kind, source, .. } => ApiError::SerializeErr {
+                kind: kind.to_string(),
                 source: format!("{}", source),
             },
-            Error::Utf8StringErr { source, .. } => ApiError::Utf8StringErr {
+            StdError::Utf8Err { source, .. } => ApiError::Utf8Err {
                 source: format!("{}", source),
             },
-            Error::Unauthorized { .. } => ApiError::Unauthorized {},
-            Error::ValidationErr { field, msg, .. } => ApiError::ValidationErr {
+            StdError::Utf8StringErr { source, .. } => ApiError::Utf8StringErr {
+                source: format!("{}", source),
+            },
+            StdError::Unauthorized { .. } => ApiError::Unauthorized {},
+            StdError::ValidationErr { field, msg, .. } => ApiError::ValidationErr {
                 field: field.to_string(),
                 msg: msg.to_string(),
             },
@@ -153,19 +153,19 @@ impl From<SystemError> for ApiSystemError {
 #[cfg(test)]
 mod test_result {
     use super::*;
-    use crate::errors::{contract_err, NoSuchContract, Result};
+    use crate::errors::{contract_err, NoSuchContract, StdResult};
     use crate::serde::{from_slice, to_vec};
 
     #[test]
     fn convert_ok_result() {
-        let input: Result<Vec<u8>> = Ok(b"foo".to_vec());
+        let input: StdResult<Vec<u8>> = Ok(b"foo".to_vec());
         let convert: ApiResult<Vec<u8>> = input.into();
         assert_eq!(convert, ApiResult::Ok(b"foo".to_vec()));
     }
 
     #[test]
     fn check_ok_into_conversion() {
-        let input: Result<bool> = Ok(true);
+        let input: StdResult<bool> = Ok(true);
         let convert: ApiResult<i32> = input.into();
         assert_eq!(convert, ApiResult::Ok(1i32));
         let expanded: Result<i64, ApiError> = convert.into();
@@ -175,7 +175,7 @@ mod test_result {
 
     #[test]
     fn convert_err_result() {
-        let input: Result<()> = contract_err("sample error");
+        let input: StdResult<()> = contract_err("sample error");
         let convert: ApiResult<()> = input.into();
         assert_eq!(
             convert,
@@ -213,7 +213,7 @@ mod test_result {
     #[test]
     // this tests Ok(Err(_)) case for SystemError, Error
     fn convert_nested_ok_err_result() {
-        let input: Result<Result<()>, SystemError> = Ok(contract_err("nested error"));
+        let input: Result<StdResult<()>, SystemError> = Ok(contract_err("nested error"));
         let convert: ApiResult<ApiResult<()>, ApiSystemError> = input.into();
         assert_eq!(
             convert,
@@ -226,7 +226,7 @@ mod test_result {
     #[test]
     // this tests Ok(Ok(_)) case for SystemError, Error
     fn convert_nested_ok_ok_result() {
-        let input: Result<Result<i32>, SystemError> = Ok(Ok(123));
+        let input: Result<StdResult<i32>, SystemError> = Ok(Ok(123));
         let convert: ApiResult<ApiResult<i32>, ApiSystemError> = input.into();
         assert_eq!(convert, ApiResult::Ok(ApiResult::Ok(123)),);
     }
@@ -234,7 +234,7 @@ mod test_result {
     #[test]
     // make sure we can shove this all over API boundaries
     fn serialize_and_recover_nested_result() {
-        let input: Result<Result<()>, SystemError> = Ok(contract_err("over ffi"));
+        let input: Result<StdResult<()>, SystemError> = Ok(contract_err("over ffi"));
         let convert: ApiResult<ApiResult<()>, ApiSystemError> = input.into();
         let recovered: ApiResult<ApiResult<(), ApiError>, ApiSystemError> =
             from_slice(&to_vec(&convert).unwrap()).unwrap();
@@ -263,11 +263,11 @@ mod test_errors {
     use super::*;
     use crate::errors::{
         contract_err, dyn_contract_err, invalid, unauthorized, Base64Err, InvalidRequest,
-        NoSuchContract, NotFound, NullPointer, Result, SerializeErr,
+        NoSuchContract, NotFound, NullPointer, SerializeErr, StdResult,
     };
     use crate::serde::{from_slice, to_vec};
 
-    fn assert_conversion(r: Result<()>) {
+    fn assert_conversion(r: StdResult<()>) {
         let error = r.unwrap_err();
         let msg = format!("{}", error);
         let converted: ApiError = error.into();