+++ packages/std/src/coins.rs
@@ -4,7 +4,7 @@ use std::convert::TryFrom;
 use std::{fmt, ops};
 
 use crate::dyn_contract_err;
-use crate::errors::Error;
+use crate::errors::StdError;
 
 #[derive(Serialize, Deserialize, Clone, Default, Debug, PartialEq, JsonSchema)]
 pub struct Coin {
@@ -56,7 +56,7 @@ impl From<u128> for Uint128 {
 }
 
 impl TryFrom<&str> for Uint128 {
-    type Error = Error;
+    type Error = StdError;
 
     fn try_from(val: &str) -> Result<Self, Self::Error> {
         match val.parse::<u128>() {
@@ -143,6 +143,7 @@ impl<'de> de::Visitor<'de> for Uint128Visitor {
 #[cfg(test)]
 mod test {
     use super::*;
+    use crate::errors::StdResult;
     use crate::{from_slice, to_vec};
     use std::convert::TryInto;
 
@@ -164,7 +165,7 @@ mod test {
         assert_eq!(34567, a.u128());
         assert_eq!("34567", a.to_string());
 
-        let a: Result<Uint128, Error> = "1.23".try_into();
+        let a: StdResult<Uint128> = "1.23".try_into();
         assert!(a.is_err());
     }