+++ packages/std/src/errors.rs
@@ -5,7 +5,10 @@ use snafu::Snafu;
 #[snafu(visibility = "pub")]
 /// Structured error type for init, handle and query. This cannot be serialized to JSON, such that
 /// it is only available within the contract and its unit tests.
-pub enum Error {
+///
+/// The prefix "Std" means "the standard error within the standard library". This is not the only
+/// result/error type in cosmwasm-std.
+pub enum StdError {
     #[snafu(display("Invalid Base64 string: {}", source))]
     Base64Err {
         source: base64::DecodeError,
@@ -62,11 +65,18 @@ pub enum Error {
     },
 }
 
+/// The return type for init, handle and query. Since the error type cannot be serialized to JSON,
+/// this is only available within the contract and its unit tests.
+///
+/// The prefix "Std" means "the standard result within the standard library". This is not the only
+/// result/error type in cosmwasm-std.
+pub type StdResult<T> = core::result::Result<T, StdError>;
+
 #[derive(Debug, Snafu)]
 #[snafu(visibility = "pub")]
 /// SystemError is used for errors inside the runtime.
 /// This is used on return values for Querier as a nested Result -
-/// Result<Result<T, Error>, SystemError>
+/// Result<StdResult<T>, SystemError>
 /// The first wrap (SystemError) will trigger if the contract address doesn't exist,
 /// the QueryRequest is malformated, etc. The second wrap will be an error message from
 /// the contract itself.
@@ -85,23 +95,19 @@ pub enum SystemError {
     Unknown { backtrace: snafu::Backtrace },
 }
 
-/// Result for init, handle and query. Since the error type cannot be serialized to JSON,
-/// this is only available within the contract and its unit tests.
-pub type Result<T, E = Error> = core::result::Result<T, E>;
-
-pub fn invalid<T>(field: &'static str, msg: &'static str) -> Result<T> {
+pub fn invalid<T>(field: &'static str, msg: &'static str) -> StdResult<T> {
     ValidationErr { field, msg }.fail()
 }
 
-pub fn contract_err<T>(msg: &'static str) -> Result<T> {
+pub fn contract_err<T>(msg: &'static str) -> StdResult<T> {
     ContractErr { msg }.fail()
 }
 
-pub fn dyn_contract_err<T>(msg: String) -> Result<T> {
+pub fn dyn_contract_err<T>(msg: String) -> StdResult<T> {
     DynContractErr { msg }.fail()
 }
 
-pub fn unauthorized<T>() -> Result<T> {
+pub fn unauthorized<T>() -> StdResult<T> {
     Unauthorized {}.fail()
 }
 
@@ -111,9 +117,9 @@ mod test {
 
     #[test]
     fn use_invalid() {
-        let e: Result<()> = invalid("demo", "not implemented");
+        let e: StdResult<()> = invalid("demo", "not implemented");
         match e {
-            Err(Error::ValidationErr { field, msg, .. }) => {
+            Err(StdError::ValidationErr { field, msg, .. }) => {
                 assert_eq!(field, "demo");
                 assert_eq!(msg, "not implemented");
             }
@@ -125,9 +131,9 @@ mod test {
     #[test]
     // example of reporting static contract errors
     fn contract_helper() {
-        let e: Result<()> = contract_err("not implemented");
+        let e: StdResult<()> = contract_err("not implemented");
         match e {
-            Err(Error::ContractErr { msg, .. }) => {
+            Err(StdError::ContractErr { msg, .. }) => {
                 assert_eq!(msg, "not implemented");
             }
             Err(e) => panic!("unexpected error, {:?}", e),
@@ -139,9 +145,9 @@ mod test {
     // example of reporting contract errors with format!
     fn dyn_contract_helper() {
         let guess = 7;
-        let e: Result<()> = dyn_contract_err(format!("{} is too low", guess));
+        let e: StdResult<()> = dyn_contract_err(format!("{} is too low", guess));
         match e {
-            Err(Error::DynContractErr { msg, .. }) => {
+            Err(StdError::DynContractErr { msg, .. }) => {
                 assert_eq!(msg, String::from("7 is too low"));
             }
             Err(e) => panic!("unexpected error, {:?}", e),